Move DirectoryHandler from searchcore to searchlib.
